Changes comparing list to HashSet and updates README
In order to increase the speed of comparison lookups for the
get_missing_paths method, this commit updates the comparing_file_names
list to be a HashSet instead of a Vector.

In addition, this commit adds a line in the README to express clarity in
how exactly the comparison_log flag works. To reduce confusion through
misconfiguration.